+++ NEWS.md
@@ -1,3 +1,7 @@
+# Rprofile (development)
+
+* Adds `.GitPrepareCommitMsg()` to copy one of two templates to `.git/hook/prepare-commit-msg` [#10]
+
 # Rprofile 0.0.0.9000
 
 * Added a `NEWS.md` file to track changes to the package.